investment-report-pydanticai/requirements.txt
pydantic==2.10.6
pydantic-ai==0.0.20
openai
pydantic-ai-slim
pydantic_core
pydantic-graph
python-dotenv==1.0.1
colorama==0.4.6
dataclasses==0.6
asyncio==3.4.3
devtools==0.12.2
html-to-markdown==1.1.0
Jinja2
yfinance
beautifulsoup4
tavily-python
asyncio